pyproject.toml
[project]
name = "fpcr"
version = "0.604.24"
description = "Firewall Policy Change Request"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "aiohttp>=3.9.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    "cp-mgmt-api-sdk>=1.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "sqlmodel>=0.0.27",
    "rich>=14.2.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "ruamel.yaml>=0.18.0",
    "cpaiops",
    "arlogi",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "cryptography>=43.0.0",
    "jsonschema>=4.0.0",
    "jinja2>=3.1.0",
    "reportlab>=4.2.0",
]

[project.optional-dependencies]
fpcr-flow = []

[tool.uv.sources]
arlogi = { git = "http://github.com/antragh/arlogi.git" }
cpaiops = { path = "./libs/cpaiops", editable = true }

[dependency-groups]
dev = [
    "mypy>=1.9.0",
    "ruff>=0.6.0",
    "pyright>=1.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.28.1",
    "types-jsonschema>=4.24.0.20251008",
    "types-reportlab>=4.2.0.20241117",
]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = false
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
mypy_path = [".", "src"]
explicit_package_bases = true

# [[tool.mypy.overrides]]
# module = ["arlogi.*", "cpaiops.*"]
# ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["*.tests", "*.tests.*"]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.13"
pythonPlatform = "Windows"
venvPath = "."
venv = ".venv"
